Refactor actions-related code
